% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GORICApower.R
\name{GORICA.power}
\alias{GORICA.power}
\title{GORICA.power}
\usage{
GORICA.power(n, model, nsim, p, B, H1, M, Omega_pop)
}
\arguments{
\item{n}{The number of variables which is represented bi- (n = 2) or tri-variate (n = 3) model}

\item{model}{Determine wave-independent (model = 1) or wave-specific parameters model (model = 2)}

\item{nsim}{The number of iterations in simulations}

\item{p}{The number of persons}

\item{B}{The matrix of population parameters}

\item{H1}{The hypothesis of interest}

\item{M}{The number of measurement occasions (The number of measurement occasions should range from 3 to 6)}

\item{Omega_pop}{The matrix of variances-covariances of the  population parameters}
}
\value{
The GORICA power and the GORICA weights
}
\description{
Calculate the GORICA power
}
\examples{
H1 <- "abs(beta6) < abs(gamma6); abs(phi6) < abs(epsilon6); abs(psi6) < abs(chi6); abs(beta5) < abs(gamma5); abs(phi5) < abs(epsilon5); abs(psi5) < abs(chi5); abs(beta4) < abs(gamma4); abs(phi4) < abs(epsilon4); abs(psi4) < abs(chi4); abs(beta3) < abs(gamma3); abs(phi3) < abs(epsilon3); abs(psi3) < abs(chi3); abs(beta2) < abs(gamma2); abs(phi2) < abs(epsilon2); abs(psi2) < abs(chi2)"
GORICA.power(n = 3, model = 2, M = 6, nsim = 3, p = 100, B = matrix(c(0.22, 0.1, 0.1, 0.4, 0.28, 0.1, 0.4, 0.4, 0.28),byrow = T, 3), H1, Omega_pop <- matrix(c(1, 0, 0, 0, 1, 0, 0, 0, 1), byrow = T, ncol = 3))
}
