% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GORICApower.R
\name{GORICA.test}
\alias{GORICA.test}
\title{GORICA.test}
\usage{
GORICA.test(n, model, M, data, H1)
}
\arguments{
\item{n}{The number of variables which is represented bi- (n = 2) or tri-variate (n = 3) model}

\item{model}{Determine wave-independent (model = 1) or wave-specific parameters model (model = 2)}

\item{M}{The number of measurement occasions (The number of measurement occasions should range from 3 to 6)}

\item{data}{The dataset}

\item{H1}{The hypothesis of interest}
}
\value{
The GORICA results, the estimates and the variances - covariances matrix of the estimates
}
\description{
Evaluate hypothesis using the GORICA
}
\examples{
require(Rfssa) # for load_github_data()

data <- load_github_data("https://github.com/Chuenjai/GORICApower3/blob/master/data3v4w.RData")

RICLPMdata <- as.data.frame(RICLPMdata)  # make sure that the dataset is in the form of data.frame

H1 <- "abs(beta4) < abs(gamma4); abs(phi4) < abs(epsilon4); abs(psi4) < abs(chi4); abs(beta3) < abs(gamma3); abs(phi3) < abs(epsilon3); abs(psi3) < abs(chi3); abs(beta2) < abs(gamma2); abs(phi2) < abs(epsilon2); abs(psi2) < abs(chi2)"

GORICA.test(n = 3, model = 2, M = 4, data, H1)
}
