% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GORICApower.R
\name{Model}
\alias{Model}
\title{Model}
\usage{
Model(n, model, M)
}
\arguments{
\item{n}{The number of variables which is represented bi- (n = 2) or tri-variate (n = 3) model}

\item{model}{Determine wave-independent (model = 1) or wave-specific parameters model (model = 2)}

\item{M}{The number of measurement occasions (The number of measurement occasions should range from 3 to 6)}
}
\value{
The review of the image of the model
}
\description{
Selecting the RI-CLPM
}
\examples{
Model(n = 3, model = 2, M = 6)

}
